src/msal/acquire_token_common_parameters.go
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT license.

package msalgo

import "github.com/AzureAD/microsoft-authentication-library-for-go/src/internal/msalbase"

type acquireTokenCommonParameters struct {
	scopes []string
}

func createAcquireTokenCommonParameters(scopes []string) *acquireTokenCommonParameters {
	p := &acquireTokenCommonParameters{
		scopes: scopes,
	}
	return p
}

func (p *acquireTokenCommonParameters) augmentAuthenticationParameters(authParams *msalbase.AuthParametersInternal) {
	authParams.SetScopes(p.scopes)
}
